move redmine_cleaner to utils + clean associated tests
